Fix rook eval (+9)
```
Score of Simbelmyne vs Simbelmyne main: 767 - 701 - 1032 [0.513]
...      Simbelmyne playing White: 490 - 246 - 514  [0.598] 1250
...      Simbelmyne playing Black: 277 - 455 - 518  [0.429] 1250
...      White vs Black: 945 - 523 - 1032  [0.584] 2500
Elo difference: 9.2 +/- 10.4, LOS: 95.8 %, DrawRatio: 41.3 %
2500 of 2500 games finished.
```

+++ simbelmyne/src/search.rs
@@ -199,7 +199,7 @@ impl From<&SearchReport> for SearchInfo {
             seldepth: Some(report.seldepth),
             time: Some(report.duration.as_millis() as u64),
             nodes: Some(report.nodes),
-            score: Some(report.score),
+            score: Some(report.score as i32),
             pv: report.pv.clone(),
             hashfull: Some(report.hashfull),
             nps: Some(nps),
